BattAnalog: Allow users to define the cell count manually
Even though the current automatic cell count detection logic works fine
for the batteries with lower cell counts, it does not work as expected
for the batteries with higher cell counts under some circumstances. For
example, 45.6V total battery voltage could either mean 4.14V per cell
for an 11S battery or 3.8V per cell for a 12S battery. In this case, the
logic in use decides that the battery is 11S and gives wrong per cell
voltage if the battery in use is 12S.

This change allows users to optionally define the cell count manually in
the settings of this widget to prevent the type of problems described
above.